Corrected ghosts returning to spawn
Ghosts will now correctly return to spawn when eaten,
and will also not interact with the avatar will in the 'killed'
state. Added a new image 'eyes.png' which the ghosts use when
in this state.

Added a function for all moving sprites 'return_to_spawn'
which returns them to their spawn location. This is used
when the avatar is hit by a ghost and loses a life, and the
positions of the avatar and ghosts is reset.